requirements.txt
absl-py==2.3.1
aniso8601==10.0.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==2.3.3
flask-cors==6.0.1
Flask-RESTful==0.3.10
flask-restx==1.1.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.42.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.13.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.7.2
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
Pillow==10.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
